+++ src/index.ts
@@ -5,6 +5,7 @@ export { default as FuzzyPicker } from './FuzzyPicker.svelte'
 export { default as Toaster } from './Toaster.svelte'
 export { type Toast, ToastQueue, toastQueue, toast } from './toast.js'
 export { type FuzzyResult, fuzzyMatch, fuzzyFilter, moveSelection } from './fuzzy.js'
+export { navigateList } from './list-navigation.js'
 
 export { default as ColorSwatch } from './ColorSwatch.svelte'
 export { default as ColorPalette } from './ColorPalette.svelte'